Dev update (#31)
* prompt upsampler system prompt is updated (#27)

* Video2world training fixes (#29)

* Video2World finetuning cleanup

* Fix training, cleanup

* CP fixes, 14B config

---------

+++ cosmos_predict2/diffusion/training/networks/general_dit_lvg.py
@@ -41,7 +41,7 @@ class VideoExtendGeneralDIT(GeneralDIT):
         del kwargs
 
         if data_type == DataType.VIDEO:
-            x_B_C_T_H_W = torch.cat([x_B_C_T_H_W, condition_video_input_mask_B_C_T_H_W], dim=1)
+            x_B_C_T_H_W = torch.cat([x_B_C_T_H_W, condition_video_input_mask_B_C_T_H_W.type_as(x_B_C_T_H_W)], dim=1)
         else:
             B, _, T, H, W = x_B_C_T_H_W.shape
             x_B_C_T_H_W = torch.cat(
